provisioning/global-device/requirements.txt
AWSIoTPythonSDK
pyOpenSSL
